Win64 compilation.
git-svn-id: http://www.bci2000.org/svn/trunk@4654 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/core/Operator/OperatorQt/MainWindow.cpp
@@ -420,7 +420,7 @@ MainWindow::UpdateDisplay()
 void
 MainWindow::SetFunctionButtons()
 {
-  for( size_t i = 0; i < Preferences::numButtons; ++i )
+  for( int i = 0; i < Preferences::numButtons; ++i )
   {
     QString name = gpPreferences->mButtons[i].Name,
             script = gpPreferences->mButtons[i].Cmd;
